# backend/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Backend
host: localhost:4000
...
basePath: /api
schemes:
  - http
  - https
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: |
     For accessing the API a valid JWT token must be passed in all the queries in
     the 'Authorization' header.
  
  
     A valid JWT token is generated by the API and retourned as answer of a call
     to the route /login giving a valid user & password.
  
  
     The following syntax must be used in the 'Authorization' header :
  
         Bearer xxxxxx.yyyyyyy.zzzzzz
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /typesOfAnswers:
    x-swagger-router-controller: typesOfAnswers.controller
    post:
      tags:
        - "Types of Answers"
      operationId: typesOfAnswersControllerPost
      summary: "Create a Type of Answer"
      description: "Create a type of Answer"
      parameters:
        - name: typeOfAnswer
          in: body
          required: true
          schema: 
            $ref: '#/definitions/typeOfAnswer'
      responses:
        "200":
          description: "Success"
        "403":
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'
    get:
      tags:
        - "Types of Answers"
      operationId: typesOfAnswersControllerGet
      summary: "Get All The Types Of Answers"
      description: "Get All The Types Of Answers"
      responses:
        "200":
          description: "Success"
        "403":
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'
  /typesOfAnswers/{idTypeOfAnswer}:
    x-swagger-router-controller: typesOfAnswers.controller
    parameters:
      - name: idTypeOfAnswer
        in: path
        description: "ID of the Type of Answer"
        required: true
        type: string
    get:
      tags:
        - "Types of Answers"
      operationId: typesOfAnswersControllerGetId
      summary: "Get Only One Type of Answer"
      description: "Get Only One Type of Answer sending the ID of Type of Answer"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: '#/definitions/typeOfAnswer'
        "403":
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'
    put:
      tags:
        - "Types of Answers"
      operationId: typesOfAnswersControllerPut
      summary: "Update Only One Type of Answer"
      description: "Update Only One Type of Answer sending the ID of Type of Answer"
      parameters:
        - name: typeOfAnswer
          in: body
          required: true
          schema: 
            $ref: '#/definitions/typeOfAnswer'
      responses:
        200:
          description: "Success"
        403:
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'
    delete:
      tags:
        - "Types of Answers"
      operationId: typesOfAnswersControllerDelete
      summary: "Delete One Type of Answer"
      description: "Delete One Type of Answer sending the ID of Type of Answer"
      responses:
        "200":
          description: "Success"
        "403":
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'
...
  /typeProcedure:
    x-swagger-router-controller: typesProcedures.controller
    post:
      tags:
        - Type Of Procedure
      operationId: typeProcedureControllerPost
      summary: "Create Type Procedure"
      description: "Create a new Type of Procedure"
      parameters:
        - name: TypeProcedure
          in: body
          required: true
          schema:
            $ref: '#/definitions/TypeProcedure'
      responses:
        200:
          description: "Success"
        403:
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'
    get:
      tags:
        - Type Of Procedure
      operationId: typeProcedureControllerGet
      summary: "Get Types of Procedures"
      description: "Get all the Type of Procedure"
      responses:
        200:
          description: "Success"
          schema:
            $ref: '#/definitions/TypeProcedure'
        403:
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'
  /typeProcedure/{idTypeProcedure}:
    x-swagger-router-controller: typesProcedures.controller
    parameters:
      - name: idTypeProcedure
        in: path
        description: "ID of Type of Type Of Procedure"
        required: true
        type: string
    get:
      tags:
        - Type Of Procedure
      operationId: typeProcedureControllerGetId
      summary: "Get Type Of Procedure"
      description: "Get only one Type Of Procedure"
      responses:
        200:
          description: "Success"
          schema:
            $ref: '#/definitions/TypeProcedure'
        403:
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'

    put:
      tags:
        - Type Of Procedure
      operationId: typeProcedureControllerPut
      summary: "Update a Type Of Procedure"
      description: "Update only one Type Of Procedure"
      parameters:
        - name: TypeProcedure
          in: body
          required: true
          schema:
            $ref: '#/definitions/TypeProcedure'
      responses:
        200:
          description: "Success"
        403:
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'
          
    delete:
      tags:
        - Type Of Procedure
      operationId: typeProcedureControllerDelete
      summary: "Delete a Type Of Procedure"
      description: "Delete only one Type Of Procedure"
      responses:
        200:
          description: "Success"
        403:
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  TypeProcedure:
    type: object
    properties:
      id:
        type: string
      typeOfProcedure:
        type: string
        example: "Cirugia"
      detail:
        type: string
        example: "Apendicectomía"
      description:
        type: string
        example: "La apendicectomía es la técnica quirúrgica por medio de la cual se extrae el apéndice (apéndice cecal, vermiforme o vermicular), habitualmente en casos de apendicitis aguda"
...
  typeOfAnswer:
    type: object
    properties:
      idTypeOfAnswer:
        type: string
      typeOfAnswer:
        type: string
        example: "TIPO DE REVISION, TIEMPO DE REVISION"
      description:
        type: string
        example: "En que instancia se encuentra para dicha revision."
      sampleValues:
        type: string
        example: "Ninguno, Leve, Moderado, Severo, Extremo"
      answers:
        type: array
        items:
          $ref: '#/definitions/answer'
      addable:
        type: boolean
        example: true
        type: string
  answer:
    type: object
    properties:
      id:
        type: string
      value:
        type: string
        example: "MODERADO"
      score:
        type: integer
        example: 2
...
  Error:
    type: object
    properties:
      message:
        type: string